add className "notranslate"
Description: add className "notranslate" to ignore the content of the code with the built-in chrome Google translator extension.
The problem is that when we translate an article into our native language, we get a mess inside the blocks with the code. Because Google-translate is trying to translate the text. The "notranslate" class causes this block to be ignored. Accordingly, it will remain unchanged.